set open status on correct time
